fixed the interface with userCardProps
