+++ src/components/userCard.tsx
@@ -2,17 +2,9 @@ import projectsCardIcon from "@/assets/icons/projectsCardIcon.svg";
 import { Avatar } from "./avatar";
 import { Card } from "./card";
 
-export interface User {
-  id: number;
-  name: string;
-  birthDate?: string;
-  position?: string;
-  salary?: number;
-  projects?: string[];
-  imageUrl?: string;
-}
+import { User } from "@/interfaces/userInterface";
 
-export interface UserCardProps {
+interface UserCardProps {
   data: User;
 }
