+++ src/interfaces/userInterface.ts
@@ -1,0 +1,9 @@
+export interface User {
+  id: number;
+  name: string;
+  birthDate?: string;
+  position?: string;
+  salary?: number;
+  projects?: string[];
+  imageUrl?: string;
+}
